requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
Django==3.2.15
django-agenda==0.7.2
django-crispy-forms==1.14.0
django-currentuser==0.5.3
django-filter==22.1
django-multiselectfield==0.1.12
django-recurrence==1.11.1
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.2.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
